+++ test/js/bun/ffi/bir/fixtures/basics/gnu-c-corner-cases-that-tcctest-exercises.c
@@ -9,12 +9,12 @@ typedef unsigned long long __attribute__((aligned(4))) packed_to_four;
          static int v1 = 34 ?: -1, v2 = 0 ?: -1;
          int old_style(a, b) int a; char *b; { return a + (b != 0); }
          static int num(int x) { return x + 1; }
-         int gnu(void) { int (*f)(int) = num; long diff; f = num + 0; diff = f - num;
+         int gnu(void) { int (*f)(int) = num; long long diff; f = num + 0; diff = f - num;
              int here = ({ __label__ l; l: 40 + 2; });
              switch (diff) { case 0: __extension__({ here++; }); }
              return v1 * 1000 + v2 * -100 + old_style((void *)3, "s") * 10000 + (f + diff)(here) * 100000; }
          void fences(int *p) { __asm__ volatile("lock; orl $0, (%%rsp)" ::: "memory"); *p = 1; __asm__ volatile("mfence"); __asm__ volatile("sfence" ::: "memory"); }
-         void stop(unsigned long why) { __asm__ volatile("int3" : : "r"(why)); __builtin_unreachable(); }
+         void stop(unsigned long long why) { __asm__ volatile("int3" : : "r"(why)); __builtin_unreachable(); }
          static inline int wide_cas(unsigned __int128 *p, unsigned __int128 *e, unsigned __int128 v) { return __atomic_compare_exchange_n(p, e, v, 0, 5, 5); }
          #define inc < dir name >
          #define dir std
